fix: ensure cancellation keymap runs only when sync is running
Previously the `<C-c>` sync cancellation keymap would be allowed to run
multiple times and print out multiple cancellation messages.

Added a check with `s.is_job_running()` to ensure that job is only
trying to be stopped if the process is running.

Otherwise, it will notify the user that the "sync job is already
exited".